// src/types/author.ts
export type AuthorDetailType = {
  id: string;
  name: string;
  biography?: string | null;
  website?: string | null;
  nationality?: string | null;
  facebook?: string | null;
  instagram?: string | null;
  xTwitter?: string | null;
  tiktok?: string | null;
  wattpad?: string | null;
  birthDate?: Date | null;
  imageUrl?: string;
};